chat_backend/requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
certifi==2022.9.24
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.1.1
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
daphne==4.0.0
defusedxml==0.7.1
Django==4.1.3
django-cors-headers==3.13.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
msgpack==1.0.4
oauthlib==3.2.2
packaging==21.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
python3-openid==3.2.0
pytz==2022.6
redis==4.3.5
requests==2.28.1
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.13
zope.interface==5.5.2
